requirements.txt
absl-py==2.1.0
appnope==0.1.4
asttokens==3.0.0
astunparse==1.6.3
beautifulsoup4==4.13.3
black==25.1.0
blinker==1.9.0
certifi==2025.1.31
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
dash==2.18.2
dash-bootstrap-components==1.7.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.13
decorator==5.2.1
distlib==0.3.9
executing==2.2.0
filelock==3.17.0
Flask==3.0.3
flatbuffers==25.2.10
fonttools==4.56.0
frozendict==2.4.6
gast==0.6.0
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.13.0
identify==2.6.8
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.8.0
kiwisolver==1.4.8
libclang==18.1.1
lxml==5.3.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mplfinance==0.12.10b0
multitasking==0.0.11
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.28.0
nest-asyncio==1.6.0
newsapi==0.1.1
newsapi-python==0.2.7
nodeenv==1.9.1
numpy==1.26.4
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathspec==0.12.1
peewee==3.17.9
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
plotly==6.0.0
pre_commit==4.1.0
prompt_toolkit==3.0.50
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.3.0
requests==2.32.3
retrying==1.3.4
rich==13.9.4
scikit-learn==1.6.1
scipy==1.15.2
setuptools==75.8.0
six==1.17.0
soupsieve==2.6
stack-data==0.6.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
tornado==6.4.2
traitlets==5.14.3
types-requests==2.32.0.20250306
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uv==0.6.6
virtualenv==20.29.2
wcwidth==0.2.13
Werkzeug==3.0.6
wheel==0.45.1
wrapt==1.17.2
xgboost==2.1.4
yfinance==0.2.54
zipp==3.21.0
